+++ tests/test_text.py
@@ -1110,3 +1110,20 @@ def test_soft_wrap() -> None:
     print(repr(output))
     expected = "\x1b[37;44m Hello World \x1b[0m\n"
     assert output == expected
+
+
+def test_soft_wrap_styled() -> None:
+    """Regression test for https://github.com/Textualize/rich/issues/3838
+
+    If soft_wrap is True and a style is set, we don't want to style the new lines.
+    """
+    console = Console(color_system="standard", width=80, force_terminal=True)
+    with console.capture() as capture:
+        console.print("soft wrap is on", style="blue on white", soft_wrap=True)
+        console.print("Next line")
+
+    output = capture.get()
+    print(repr(output))
+    # Background is reset before \n
+    expected = "\x1b[34;47msoft wrap is on\x1b[0m\nNext line\n"
+    assert output == expected